Update tags and prompt
